// app/insurance/insurance_sold/insurance_sold.interface.ts
export interface InsuranceSold {

    numberProposal: String;
    proposal: String;
    adp: String;
    clientType: String;
    name: String;
    statusSeguro: String;
    saleTypeId: String;
    dateCreationInit: Date;
    dateCreationEnd: Date;
    dealership: String;
    salesMan: String;
    
}